Fix: lttng: add-trigger: payload capture fields specified as contexts
ir_op_load_expr_to_event_expr() re-uses the same `load_expr_op` variable
to traverse the IR. A check for the type of the original load expression
is performed after `load_expr_op->next` is assigned to load_expr_op
which is expected to have remained the same.

The original event_expr child type is sampled at the start of the
function and re-used as needed.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I14ebcb9088c5107f3133ff5b43c65c3fc6101077