clock override: use getter API from LTTng-UST
Currently, the lttng session daemon reimplements its own copy of
struct lttng_trace_clock, which is pretty bad as ABI go.

Use getter functions for each of the clock callback instead. This
refactoring needs to be introduced with a matching commit to implement
the new ABI in LTTng-UST.

Move the implementation of the clock frequency, uuid, name and
description default callbacks (for monotonic clock) to lttng-ust.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: If2e671bacacc55d3fb27b6b4d4fc8c16aaffd059