Fix: test: start-stop trigger: test execution is invalid
Observed issue
==============

The test expects the notification client to be a background process but
`&` is not used.

The notification client is expected to sync with its launcher via the
sync file but the invocation of mktemp already creates the file.

Solution
=======

Use `&` and `mktemp -u`.

While there a little cleanup in notification_client for error code path
was done.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Iba80413d6ca36989cd967895d3330860cb4ef614